# api-gateway/src/main/resources/static/mobile-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YeYamo Mobile Gateway Surface",
    "version": "1.0.0-local",
    "description": "Index OpenAPI versionné de la surface mobile. Les OpenAPI générées de chaque microservice restent la source des DTO et validations détaillés."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8083",
      "description": "Même machine / Expo Web"
    },
    {
      "url": "http://10.0.2.2:8083",
      "description": "Android Emulator"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/users/social/settings": {
      "get": {
        "operationId": "getSocialSettings",
        "summary": "getSocialSettings",
        "x-yeyamo-service": "user-service",
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      },
      "put": {
        "operationId": "updateSocialSettings",
        "summary": "updateSocialSettings",
        "x-yeyamo-service": "user-service",
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
    "/api/v1/users/social/following": {
      "get": {
        "operationId": "following",
        "summary": "following",
        "x-yeyamo-service": "user-service",
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
    "/api/v1/users/social/followers": {
      "get": {
        "operationId": "followers",
        "summary": "followers",
        "x-yeyamo-service": "user-service",
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
    "/api/v1/users/social/{userId}/follow": {
      "post": {
        "operationId": "follow",
        "summary": "follow",
        "x-yeyamo-service": "user-service",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      },
      "delete": {
        "operationId": "unfollow",
        "summary": "unfollow",
        "x-yeyamo-service": "user-service",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
    "/api/v1/users/social/blocked": {
      "get": {
        "operationId": "blocked",
        "summary": "blocked",
        "x-yeyamo-service": "user-service",
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
    "/api/v1/users/social/followers/{userId}": {
      "delete": {
        "operationId": "removeFollower",
        "summary": "removeFollower",
        "x-yeyamo-service": "user-service",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "notifications",
        "summary": "notifications",
        "x-yeyamo-service": "notification-service",
        "responses": {
          "200": {
            "description": "Succès. Consulter le schéma généré du service propriétaire pour le DTO complet."
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "JWT absent ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Accès interdit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ressource absente ou masquée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit métier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Limite de débit dépassée"
          },
          "503": {
            "description": "Service indisponible"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": [
              "integer",
              "null"
            ]
          },
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "object"
            }
          },
          "timestamp": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "correlationId": {
            "type": [
              "string",
              "null"
            ]
          },
          "retryable": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "retryAfter": {
            "type": [
              "integer",
              "null"
            ]
          }
        }
      }
    }
  }
}